Moved SpaCy components and their packages into a docproc package
