+++ src/conspiracies/__init__.py
@@ -1,2 +1,2 @@
 from .registry import registry  # noqa F401
-from .doc_utils import docs_from_jsonl, docs_to_jsonl  # noqa F401
+from conspiracies.docproc.doc_utils import docs_from_jsonl, docs_to_jsonl  # noqa F401
