+++ src/conspiracies/docproc/relationextraction/gptprompting/data_classes.py
