+++ tests/test_prompt_data_classes.py
@@ -1,6 +1,6 @@
 import pytest
 import spacy
-from conspiracies.relationextraction.gptprompting import (
+from conspiracies.docproc.relationextraction.gptprompting import (
     DocTriplets,
     SpanTriplet,
     StringTriplet,
